+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fixed the farmer not funneling seeds back to the seed slots.
 * Added crops block-tag to Canola/Flax/Rice
 * Fixed the drill block placement augment.
+* Migrated the Coal generator to using any burnable item, Removed Solid fuel recipes.
 
 # 1.3.11+mc1.21.1
 * Fixed Farmer not playing well with non-vanilla farmland.